LayoutTests/http/tests/security/cross-frame-access-parent-explicit-domain-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/resources/cross-frame-iframe.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-parent-explicit-domain.html. The frame requesting access set 'document.domain' to '127.0.0.1', but the frame being accessed did not. Both must set 'document.domain' to the same value to allow access.


PASS: Cross frame access from frame explicitly setting document.domain was denied.


--------
Frame: 'aFrame'
--------
Inner iframe.
